.
Added ability for contacts to have captions, that describe them.

Added possibility for a user on the public web to update his notification contacts, or an admin changing the notification contacts of a user.

Added possibility of frontend components to know the information to provide in order to create a contact for a particular provider. This information is called the contactForm.