.
Fixed minor bug with the contact-input widget, which occurs when there are no registered contact providers in the system.

Added checkOnboardingStatus(), and checkMyOnboarding() methods which other components can call to check if a user completed onboarding.

Fixed minor bug with token-based authentication.

Made it possible for a value to be passed to the zone-input widget.